[feat/refactor] raceModel 로직 옮기기
